% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filings10K2015}
\alias{filings10K2015}
\title{All FTP links to filed 10-K reports in 2015}
\format{data frame, 5671 rows, 5 variables:
\describe{
  \item{CIK}{Central Index Key (CIK) code for each company. See \url{https://www.sec.gov/edgar/searchedgar/cik.htm}.}
  \item{company_name}{name of the company.}
  \item{report_type}{Type of filing. Should all be 10-K files.}
  \item{filing_date}{Date the 10-K report was filed to the SEC.}
  \item{ftp_url}{URL pointing to the report hosted on the SEC FTP server.}
}}
\source{
\url{ftp://ftp.sec.gov/edgar/full-index/2014/QTR1/master.idx}
}
\usage{
filings10K2015
}
\description{
A dataset containing FTP links to all filed 10-K reports in 2015.
}
\keyword{datasets}
