% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filingsByYear.R
\name{filingsByYear}
\alias{filingsByYear}
\title{Retrieve all submitted 10-K filings for a particular year}
\usage{
filingsByYear(year)
}
\arguments{
\item{year}{Integer. Input cannot be less than 1993 or higher than the previous year.}
}
\value{
Returns a data frame of dimensions n x k containing all 10-K filings for that year.
}
\description{
\code{filingsByYear} retrieves all the urls of 10-K filings for a particular year. (e.g. \url{https://www.sec.gov/Archives/edgar/full-index/2004/}). It downloads the 'master.idx' file for each quarter and filters out the 10-K filings. Note that a company may submit multiple 10-K reports. These will differ by their ARC number. Also note that, prior to 2004, the SEC does not provide HTML versions of 10-K reports. Consequently, the TenK package will not be able to retrieve content prior to 2004.
}
\seealso{
See \url{https://goo.gl/pNMMcK} for documentation on this package.
  See also \code{\link{TenK_process}}.
}
\author{
Jasper Ginn
}
