% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TenK_get.R
\name{TenK_get}
\alias{TenK_get}
\title{Converts an FTP url from an SEC 10-K report to a URL pointing towards the
HTML version of that report.}
\usage{
TenK_get(URL, type = c("ftp", "html"), meta_list)
}
\arguments{
\item{URL}{url belonging to the 10-k annual report}

\item{type}{Is the url an FTP or an HTML url?}

\item{meta_list}{Function returns a list with additional meta information
(See overview of variables here: \url{https://goo.gl/AJFUul}). Choose which
fields you want to return. Defaults to "ALL".}
}
\value{
Depending on the 'meta_list' parameter, the function returns a list
  with useful information. Please refer to the documentation for additional
  information. To view the documentation, execute 'vignette("TenK")'
}
\description{
\code{TenK_get} converts an FTP url from an SEC 10-K report to a URL pointing
towards the HTML version of that report. The function will also collect and
return metadata if desired by the user. This function should not be called
directly; it is called from \code{TenK_process}.
}
\seealso{
See \url{https://goo.gl/pNMMcK} for documentation on this package.
  See also \code{\link{TenK_process}}.
}
\author{
Jasper Ginn
}
