% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TenK_process.R
\name{TenK_process}
\alias{TenK_process}
\title{Scrape an entire 10-K report or the business description section/}
\usage{
TenK_process(URL, metadata = TRUE, meta_list = list(), retrieve = c("ALL",
  "BD"))
}
\arguments{
\item{URL}{url belonging to the 10-k annual report}

\item{metadata}{If FALSE, the function does not return any metadata. In this
case, it will only return the url to the HTML 10-K report and the 10-K
document or business description.}

\item{meta_list}{Choose which fields you want to return. Defaults to "ALL".
For an overview of the variables that you can select, please see
\url{https://goo.gl/AJFUul}.}

\item{return}{Return either full report ("ALL") or business description ("BD").
Defaults to "ALL"}
}
\value{
Depending on the 'return_meta' parameter and 'meta_list' parameter,
  the function returns a list with useful information. Please refer to the
  documentation for additional information.
}
\description{
\code{TenK_process} retrieves the entire 10-K annual report or the business
  description from an url belonging to a SEC 10-K report. If an FTP url is
  passed, the function will determine the link to the HTML version. See
  documentation (execute 'vignette("TenK")' or visit \url{https://goo.gl/pNMMcK})
}
\examples{
res <- TenK_process("ftp://ftp.sec.gov/edgar/data/34782/0000034782-16-000102.txt")

}
\author{
Jasper Ginn
}
\seealso{
See \url{https://goo.gl/pNMMcK} for documentation on this package. You can
  also execute 'vignette("TenK")'.
}

